+++ src/widgets/dialog.c
@@ -22,6 +22,7 @@
 #include "tkc/mem.h"
 #include "base/idle.h"
 #include "widgets/label.h"
+#include "widgets/button.h"
 #include "base/enums.h"
 #include "tkc/utils.h"
 #include "base/layout.h"
@@ -151,11 +152,19 @@ widget_t* dialog_get_client(widget_t* widget) {
 
 #include "base/window_manager.h"
 
-#define TOAST_STYLE_NAME "default"
+#define OK_STYLE_NAME "default"
+#define CANCEL_STYLE_NAME "default"
 
-static ret_t on_event_to_quit(void* ctx, event_t* e) {
+static ret_t on_ok_to_quit(void* ctx, event_t* e) {
   widget_t* dlg = WIDGET(ctx);
-  dialog_quit(dlg, 0);
+  dialog_quit(dlg, RET_OK);
+
+  return RET_REMOVE;
+}
+
+static ret_t on_cancel_to_quit(void* ctx, event_t* e) {
+  widget_t* dlg = WIDGET(ctx);
+  dialog_quit(dlg, RET_FAIL);
 
   return RET_REMOVE;
 }
@@ -170,45 +179,164 @@ static ret_t on_timer_to_quit(const timer_info_t* info) {
 
 /*simple dialogs*/
 
-ret_t dialog_toast(const char* text, uint32_t duration) {
-  xy_t x = 0;
-  xy_t y = 0;
+static ret_t dialog_get_size_limits(rect_t* r) {
   uint32_t min_w = 128;
   uint32_t max_w = 0;
   uint32_t min_h = 30;
   uint32_t max_h = 0;
-  widget_t* wm = NULL;
+  widget_t* wm = window_manager();
+
+  max_w = wm->w * 0.8;
+  max_h = wm->h * 0.8;
+
+  r->x = min_w;
+  r->w = max_w;
+  r->y = min_h;
+  r->h = max_h;
+
+  return RET_OK;
+}
+
+static widget_t* dialog_create_label(const char* text) {
+  rect_t r;
+  widget_t* label = NULL;
+
+  dialog_get_size_limits(&r);
+
+  label = label_create(NULL, 0, 0, 0, 0);
+  return_value_if_fail(label != NULL, NULL);
+
+  widget_set_text_utf8(label, text);
+  label_resize_to_content(label, r.x, r.w, r.y, r.h);
+
+  return label;
+}
+
+ret_t dialog_toast(const char* text, uint32_t duration) {
   widget_t* label = NULL;
   widget_t* dialog = NULL;
   return_value_if_fail(text != NULL, RET_BAD_PARAMS);
 
-  dialog = dialog_create(NULL, 0, 0, 200, 30);
-  return_value_if_fail(dialog != NULL, RET_OOM);
-  label = label_create(dialog, 0, 0, dialog->w, dialog->h);
-  goto_error_if_fail(label != NULL);
+  label = dialog_create_label(text);
+  return_value_if_fail(label != NULL, RET_OOM);
 
-  wm = dialog->parent;
+  dialog = dialog_create(NULL, 0, 0, label->w, label->h);
+  goto_error_if_fail(dialog != NULL);
 
-  max_w = wm->w * 0.8;
-  max_h = wm->h * 0.8;
-  widget_set_text_utf8(label, text);
-  widget_use_style(label, TOAST_STYLE_NAME);
-  widget_use_style(dialog, TOAST_STYLE_NAME);
-  label_resize_to_content(label, min_w, max_w, min_h, max_h);
+  widget_set_prop_str(dialog, WIDGET_PROP_ANIM_HINT, "fade(duration=500)");
+  widget_set_prop_str(dialog, WIDGET_PROP_THEME, "dialog_toast");
+  widget_add_child(dialog, label);
 
-  x = (wm->w - label->w) >> 1;
-  y = (wm->h - label->h) >> 1;
-  widget_resize(dialog, label->w, label->h);
-  widget_move(dialog, x, y);
-
-  widget_on(dialog, EVT_POINTER_UP, on_event_to_quit, dialog);
+  widget_on(dialog, EVT_POINTER_UP, on_ok_to_quit, dialog);
   widget_add_timer(dialog, on_timer_to_quit, duration);
 
-  dialog_modal(dialog);
-
-  return RET_OK;
+  return dialog_modal(dialog);
 error:
-  widget_destroy(dialog);
+  widget_destroy(label);
 
   return RET_FAIL;
+}
+
+ret_t dialog_info_ex(const char* text, const char* title_text, const char* theme) {
+  uint32_t w = 0;
+  uint32_t h = 0;
+  char params[128];
+  widget_t* ok = NULL;
+  widget_t* label = NULL;
+  widget_t* dialog = NULL;
+  widget_t* title = NULL;
+  widget_t* client = NULL;
+  return_value_if_fail(text != NULL, RET_BAD_PARAMS);
+
+  label = dialog_create_label(text);
+  return_value_if_fail(label != NULL, RET_OOM);
+
+  tk_snprintf(params, sizeof(params) - 1, "default(x=10, y=10, w=%d, h=%d)", label->w, label->h);
+  widget_set_self_layout(label, params);
+
+  h = label->h + 90;
+  w = tk_max(label->w, 128) + 20;
+  dialog = dialog_create_simple(NULL, 0, 0, w, h);
+  widget_set_prop_str(dialog, WIDGET_PROP_THEME, theme);
+
+  client = dialog_get_client(dialog);
+  title = dialog_get_title(dialog);
+
+  widget_set_tr_text(title, title_text);
+  widget_add_child(client, label);
+
+  ok = button_create(client, 0, 0, 0, 0);
+  widget_set_tr_text(ok, "OK");
+  widget_set_self_layout(ok, "default(x=c, y=bottom:10, w=50%, h=30)");
+  widget_on(ok, EVT_CLICK, on_ok_to_quit, dialog);
+
+  if (dialog != NULL && ok != NULL) {
+    widget_layout(dialog);
+
+    return dialog_modal(dialog);
+  } else {
+    widget_destroy(dialog);
+
+    return RET_FAIL;
+  }
+}
+
+ret_t dialog_info(const char* text) {
+  return dialog_info_ex(text, "Infomation", "dialog_info");
+}
+
+ret_t dialog_warn(const char* text) {
+  return dialog_info_ex(text, "Warnning", "dialog_warn");
+}
+
+ret_t dialog_confirm(const char* text) {
+  uint32_t w = 0;
+  uint32_t h = 0;
+  char params[128];
+  widget_t* ok = NULL;
+  widget_t* cancel = NULL;
+  widget_t* label = NULL;
+  widget_t* dialog = NULL;
+  widget_t* title = NULL;
+  widget_t* client = NULL;
+  return_value_if_fail(text != NULL, RET_BAD_PARAMS);
+
+  label = dialog_create_label(text);
+  return_value_if_fail(label != NULL, RET_OOM);
+
+  tk_snprintf(params, sizeof(params) - 1, "default(x=10, y=10, w=%d, h=%d)", label->w, label->h);
+  widget_set_self_layout(label, params);
+
+  h = label->h + 90;
+  w = tk_max(label->w, 128) + 20;
+  dialog = dialog_create_simple(NULL, 0, 0, w, h);
+  widget_set_prop_str(dialog, WIDGET_PROP_THEME, "dialog_confirm");
+
+  client = dialog_get_client(dialog);
+  title = dialog_get_title(dialog);
+
+  widget_set_tr_text(title, "Confirm");
+  widget_add_child(client, label);
+
+  ok = button_create(client, 0, 0, 0, 0);
+  widget_set_tr_text(ok, "OK");
+  widget_use_style(ok, OK_STYLE_NAME);
+  widget_set_self_layout(ok, "default(x=10%, y=bottom:10, w=30%, h=30)");
+  widget_on(ok, EVT_CLICK, on_ok_to_quit, dialog);
+
+  cancel = button_create(client, 0, 0, 0, 0);
+  widget_set_tr_text(cancel, "Cancel");
+  widget_use_style(cancel, CANCEL_STYLE_NAME);
+  widget_set_self_layout(cancel, "default(x=r:10%, y=bottom:10, w=30%, h=30)");
+  widget_on(cancel, EVT_CLICK, on_cancel_to_quit, dialog);
+
+  if (dialog != NULL && ok != NULL) {
+    widget_layout(dialog);
+
+    return dialog_modal(dialog);
+  } else {
+    widget_destroy(dialog);
+
+    return RET_FAIL;
+  }
 }